ai_api/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["daria425 <darianaumova5@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.8.0"}
python-dotenv = "^1.0.1"
ipython = "^8.26.0"
google-cloud-aiplatform = "^1.57.0"
google-cloud-translate = "^3.15.3"
beautifulsoup4 = "^4.12.3"
fastapi = "^0.111.0"
gunicorn = "^22.0.0"
twilio = "^9.3.2"
google-cloud-speech = "^2.27.0"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
